feat(DatePicker): add FilterChipDatePicker component [FC- 3341] (#2559)
* feat: add BaseFilterChip

* feat: filter chip selector input

* chore: update storie

* chore: added more exmaple

* feat: added more example and bug fix

* chore: remove todo

* chore: added whitespace

* chore: more type fixes

* chore: add changeset

* chore: rename

* chore: update snaps

* feat: initial implmentation

* feat: added single date selection

* chore: change input

* chore: added selection type multiple

* feat: add hanle reset

* chore: change label

* fix: few type errors

* chore: remove partial

* chore: added types

* chore: remove type changes

* refactor: datepicker logic

* chore: add id in base filter

* chore: make id optional

* fix: added react native

* chore: added example with presets

* fix: date selection bug

* chore: change selection type

* chore: remove unwanted warnings

* chore: remove unused console.log

* refactor: change datepicker to datepicker wrapper

* fix: more errors

* chore: remove stories

* chore: remove comments

* chore: type change for datepicker filter

* chore: fix dublicate issue

* chore: add changeset

* fix: ts error

* fix: base filter chip

* chore: update base filter chip

* chore: fix ts error

* chore: import issue

* fix: lint

* fix: linting errors

* chore: change name to filterchip prop

* chore: minor review changes

* chore: update change log

* chore: added controlled example

* chore: update changelog

* chore: review changes

* feat: added uncontrolled handling for filterChipSelect Input

* fix: typecheck

* chore: remove key

* chore: update snap

* chore: update snaps and validate error fix

* chore: update readme.md

* chore: update changeset

* chore: sync basedatepicker with main datepicker

* fix: make sure to use web calendar footer

* chore: remove changeset

---------
