+++ .changeset/hot-grapes-care.md
@@ -2,4 +2,4 @@
 '@razorpay/blade': minor
 ---
 
-feat(blade): add FilterChipSelectInput
+feat(blade): add `FilterChipSelectInput` & `FilterChipDatePicker` components
