+++ packages/blade/src/components/DatePicker/FilterChipDatePicker/FilterChipDatePicker.native.tsx
@@ -1,0 +1,13 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import { throwBladeError } from '~utils/logger';
+
+const FilterChipDatePicker = (_props: any): React.ReactElement => {
+  throwBladeError({
+    message: 'FilterChipDatePicker is not yet implemented for native',
+    moduleName: 'FilterChipDatePicker',
+  });
+
+  return <></>;
+};
+
+export { FilterChipDatePicker };
