+++ packages/blade/src/components/DatePicker/FilterChipDatePicker/index.ts
@@ -1,0 +1,1 @@
+export { FilterChipDatePicker } from './FilterChipDatePicker';
