+++ packages/blade/src/components/DatePicker/index.ts
@@ -1,2 +1,3 @@
 export * from './DatePicker';
+export { FilterChipDatePicker } from './FilterChipDatePicker';
 export type { DatePickerProps, DateValue, DatesRangeValue } from './types';
