+++ packages/blade/src/components/DatePicker/useDatesState.ts
@@ -100,6 +100,11 @@ export function useDatesState<Type extends DatePickerType = 'default'>({
     return false;
   };
 
+  const handleReset = () => {
+    setPickedDate(null);
+    setHoveredDate(null);
+    setValue([null, null]);
+  };
   const onRootMouseLeave =
     type === 'range'
       ? (event: React.MouseEvent<HTMLDivElement>) => {
@@ -173,5 +178,6 @@ export function useDatesState<Type extends DatePickerType = 'default'>({
     pickedDate,
     controlledValue: _value,
     setControlledValue: setValue,
+    handleReset,
   } as const;
 }
